Correct error of J in model_func

+++ Floquet/model_func.cpp
@@ -10,7 +10,7 @@ void FloEvolRandomFunc::operator() (const AllPara& parameters,
 EvolMatrix< ComplexEigenSolver<MatrixXcd> >*& model){
 	const int size = parameters.generic.size; // System Size
 	const double tau = parameters.floquet.tau; // Time step, which seems not used here
-	const int J = parameters.floquet.J; // Coupling strength
+	const double J = parameters.floquet.J; // Coupling strength
 
 	const bool debug = parameters.generic.debug;
 
@@ -29,7 +29,7 @@ void FloEvolRandomRotationFunc::operator() (const AllPara& parameters,
 EvolMatrix< ComplexEigenSolver<MatrixXcd> >*& model){
 	const int size = parameters.generic.size; // System Size
 	const double tau = parameters.floquet.tau; // Time step, which seems not used here
-	const int J = parameters.floquet.J; // Coupling strength
+	const double J = parameters.floquet.J; // Coupling strength
 
 	const double angle_min = parameters.floquet_random.angle_min; // Minimum angle
 	const double angle_sup = parameters.floquet_random.angle_sup; // Supreme angle
